requirements.txt
asgiref==3.9.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cryptography==45.0.6
Django==5.2.5
django-cors-headers==4.7.0
django-filter==25.1
django-oauth-toolkit==3.0.1
djangorestframework==3.16.1
gunicorn==23.0.0
idna==3.10
jwcrypto==1.5.6
oauthlib==3.3.1
packaging==25.0
psycopg2-binary==2.9.10
pycparser==2.22
python-decouple==3.8
requests==2.32.4
requests-oauthlib==2.0.0
sqlparse==0.5.3
typing_extensions==4.14.1
urllib3==2.5.0
